+++ sample-code/spring-app/src/main/java/com/sap/ai/sdk/app/services/OrchestrationService.java
@@ -120,8 +120,7 @@ public class OrchestrationService {
   @Nonnull
   public OrchestrationChatResponse template(@Nonnull final String language) {
     val template = Message.user("Reply with 'Orchestration Service is working!' in {{?language}}");
-    val templatingConfig =
-        TemplateConfig.create().withTemplate(List.of(template.createChatMessage()));
+    val templatingConfig = TemplateConfig.create().withTemplateMessages(List.of(template));
     val configWithTemplate = config.withTemplateConfig(templatingConfig);
 
     val inputParams = Map.of("language", language);
@@ -428,9 +427,7 @@ public class OrchestrationService {
   public OrchestrationChatResponse responseFormatJsonObject(@Nonnull final String word) {
     val template = Message.user("What is '%s' in German?".formatted(word));
     val templatingConfig =
-        TemplateConfig.create()
-            .withTemplate(List.of(template.createChatMessage()))
-            .withJsonResponse();
+        TemplateConfig.create().withTemplateMessages(List.of(template)).withJsonResponse();
     val configWithTemplate = config.withTemplateConfig(templatingConfig);
 
     val prompt =